+++ client/src/components/ResusableComponents/Posting.jsx
@@ -2,6 +2,8 @@ import { PencilIcon, TrashIcon } from "@heroicons/react/solid";
 import React, { useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import More_Icon from "../../assets/More_Icon";
+
+
 const Posting = ({ posting }) => {
   const [open, setopen] = useState(false);
 
@@ -23,7 +25,11 @@ const Posting = ({ posting }) => {
             {posting.description}
           </p>
           <button className="font-poppins rounded mt-5 mb-5 mr-2 bg-blue-500 hover:bg-blue-700 text-white font-medium py-2 px-4 focus:outline-none focus:shadow-outline">
-            <PencilIcon className="w-4 h-4 text-white" aria-hidden="true" onClick={() => navigate("inf")}/>
+            <PencilIcon
+              className="w-4 h-4 text-white"
+              aria-hidden="true"
+              onClick={() => navigate("inf")}
+            />
           </button>
           <button className="font-poppins mt-5 mb-5 bg-blue-500 hover:bg-blue-700 text-white font-medium py-2 px-4 rounded focus:outline-none focus:shadow-outline">
             <TrashIcon className="w-4 h-4 text-white" aria-hidden="true" />
